Spell Refactoring done
Easier to expand spell functionality and create new Effects and types

+++ Project/Assets/Assets/Scripts/Interfaces/ICharacter.cs
@@ -19,8 +19,8 @@ public interface ICharacter
     // *----- Attributes -----*
     float CharacterMovementSpeed { get; }
     float CharacterAttackSpeed { get; }
-    int CharacterAttackPower { get; }
-    int CharacterDefense { get; }
+    float CharacterAttackPower { get; }
+    float CharacterDefense { get; }
     List<Resistance> CharacterResistances { get; }
 
     // *----- Equipables and Inventory -----*
